set type of trigger cell to unit16_t and actually set it to the waveform class instance

+++ main/include/eudaq/StandardEvent.hh
@@ -28,7 +28,7 @@ public:
 	} //todo: FIx issue with template
 //	void SetWaveform(float* data);
 	std::vector<float>* GetData() const{return &m_samples;};
-	void SetTriggerCell(unsigned trigger_cell){m_trigger_cell=trigger_cell;}
+	void SetTriggerCell(uint16_t trigger_cell){m_trigger_cell=trigger_cell;}
 	uint16_t GetTriggerCell() const{return m_trigger_cell;}
 	unsigned ID() const;
 	void Print(std::ostream &) const;
@@ -80,7 +80,7 @@ private:
 	int m_channelnumber;
 	mutable std::vector<float> m_samples;
 	unsigned m_id;
-	std::string m_type, m_sensor,m_channelname;
+	std::string m_type, m_sensor, m_channelname;
 	uint16_t m_trigger_cell;
 
 };
